Add Fabric loader project with shared gametests
fabric/ implements the platform seam: immediate registry helper, vanilla
tool-action maps (via access widener), Fabric API events for
death/damage/ticks/commands/creative tab/enchant gating, LootTableEvents
pools replacing the NeoForge GLM, Forge Config API Port (bundled
jar-in-jar) for the same TOML config, and three small mixins for the
hooks Fabric API lacks (break speed, pre-damage armor traits, armor
durability skip). Gametest bodies moved to common/GameTestBodies and run
on both loaders (26 NeoForge / 24 Fabric, all passing).
