+++ fabric/src/main/resources/META-INF/services/dev.marston.randomloot.platform.services.RegHelper
@@ -1,0 +1,1 @@
+dev.marston.randomloot.fabric.FabricRegHelper
