+++ fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,38 @@
+{
+    "schemaVersion": 1,
+    "id": "${mod_id}",
+    "version": "${mod_version}",
+    "name": "${mod_name}",
+    "description": "${mod_description}",
+    "authors": [
+        "${mod_authors}"
+    ],
+    "contact": {
+        "homepage": "https://github.com/TheMarstonConnell/randomloot",
+        "sources": "https://github.com/TheMarstonConnell/randomloot"
+    },
+    "license": "${mod_license}",
+    "icon": "logo.png",
+    "environment": "*",
+    "entrypoints": {
+        "main": [
+            "dev.marston.randomloot.fabric.RandomLootFabric"
+        ],
+        "client": [
+            "dev.marston.randomloot.fabric.RandomLootFabricClient"
+        ],
+        "fabric-gametest": [
+            "dev.marston.randomloot.fabric.RandomLootFabricGameTests"
+        ]
+    },
+    "mixins": [
+        "${mod_id}.fabric.mixins.json"
+    ],
+    "accessWidener": "${mod_id}.accesswidener",
+    "depends": {
+        "fabricloader": ">=${fabric_loader_version}",
+        "fabric-api": "*",
+        "minecraft": "~${minecraft_version}",
+        "java": ">=25"
+    }
+}
