+++ fabric/src/main/resources/randomloot.accesswidener
@@ -1,0 +1,6 @@
+accessWidener v2 official
+# randomloot:cosmetic item model property registration (NeoForge exposes this via an event)
+accessible field net/minecraft/client/renderer/item/properties/numeric/RangeSelectItemModelProperties ID_MAPPER Lnet/minecraft/util/ExtraCodecs$LateBoundIdMapper;
+# vanilla tool-action conversion maps, used by FabricPlatformHelper.getToolModifiedState
+accessible field net/minecraft/world/item/AxeItem STRIPPABLES Ljava/util/Map;
+accessible field net/minecraft/world/item/ShovelItem FLATTENABLES Ljava/util/Map;
