+++ fabric/src/main/resources/randomloot.fabric.mixins.json
@@ -1,0 +1,14 @@
+{
+    "required": true,
+    "minVersion": "0.8",
+    "package": "dev.marston.randomloot.fabric.mixin",
+    "compatibilityLevel": "JAVA_25",
+    "mixins": [
+        "ItemStackMixin",
+        "LivingEntityMixin",
+        "PlayerMixin"
+    ],
+    "injectors": {
+        "defaultRequire": 1
+    }
+}
