+++ gradle.properties
@@ -8,11 +8,11 @@ org.gradle.configuration-cache=true
 # Environment Properties
 # You can find the latest versions here: https://projects.neoforged.net/neoforged/neoforge
 # The Minecraft version must agree with the Neo version to get a valid artifact
-minecraft_version=26.2.0
+minecraft_version=26.2
 # The Minecraft version range can use any release version of Minecraft as bounds.
 # Snapshots, pre-releases, and release candidates are not guaranteed to sort properly
 # as they do not follow standard versioning conventions.
-minecraft_version_range=[26.2.0]
+minecraft_version_range=[26.2]
 # The Neo version must agree with the Minecraft version to get a valid artifact
 neo_version=26.2.0.11-beta
 # The Neo version range can use any version of Neo as bounds
